my-app/backend/requirements.txt
google-generativeai
python-dotenv
Flask
flask-cors
beautifulsoup4
